fixing reviewer recommendations
